social_blogging_app/requirements.txt
uvicorn
crewai
langchain
langchain-google-genai
google-generativeai
chromadb
python-dotenv